Angles are floats, not ints :) Also uncommented the port initialization, preliminary tests are looking good!

+++ real_time/surface_src/java_src/Alice/src/main/java/com/roboclub/robobuggy/nodes/RBSMNode.java
@@ -158,7 +158,7 @@ public class RBSMNode extends SerialNode implements Node {
 		
 		if(sensorName.equals("steering")) {
 			data.put("name", "Steering");
-			params.put("angle", Integer.valueOf(messageData[2]));
+			params.put("angle", Float.valueOf(messageData[2]));
 		}
 		else if (sensorName.equals("encoder")) {
 			data.put("name", "Encoder");
